+++ lib/extensions/bool_check.dart
@@ -1,0 +1,5 @@
+extension boolExt on bool? {
+  /// Verifica valor booleano de variável que pode ser nula.
+  /// Considera `null` equivalente à `false`.
+  bool get check => this ?? false;
+}